Added Bias, Sigmoid and Checked Off Params
I spent a good deal of time today comparing the Smart Sweeper results to
my results. The height of my fitness curve is significantly stunted, but
the rates seem similar. I'm not sure what the source of the amplitude
difference is but the only significant different feature I've found so
far is that I do not implement Elitism in the genetic algorithm.