Almost There!
This version is much cleaner and nicer than previous version. The only
additional major change that should be made is that the "Find Nearest
Food" function (and the associated direction calculation) should be
updated to take into account the toroidal nature of the map.

I'm guessing this will have a deglegable effect.